設定ファイルダウンロード機能をコンポーネント化
